fix(frontend): sessionLoading no longer gates send button; refactor(backend): extract AgentCLI interface; chore: housekeeping
Closes #673 — remove || sessionLoading from RepositoryPage disabled conditions
  (AgentSelector, Model select, Send button). sessionLoading is visual-only;
  chatAgentProcessing is the correct send guard. Remove stale JSDoc bullet from
  useSessionLoader.ts. Update RepositoryPage test.

Closes #705 (partial: useAgentPolling + ChatWindow) — remove stale JSDoc line
  referencing RepositoryPage.tsx:1249-1275; extract ChatFooter component from
  ChatWindow.tsx to eliminate the duplicated spinner + session-ID JSX block.

Closes #680 — move OpenCodeAgentCLI from agent_cli.py into new
  opencode_legacy_agent_cli.py. agent_cli.py is now interface-only (225 lines).
  Remove duplicate _parse_export_messages, _extract_part_content, and
  _to_milliseconds definitions from the legacy class. Update all imports.

Note: #676 was already fixed (useApi.ts:22-25) — closed without code change.
